Size type uint32_t replaced with uint64_t

+++ include/aspas_merge_avx.tcc
@@ -758,17 +758,17 @@ in_register_merge_key(T& v0, T& v1, __m256i &pv0, __m256i &pv1)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m256i vec0;
     __m256i vec1;
     
     uint8_t stride = (uint8_t)simd_width::AVX_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -937,7 +937,7 @@ merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
 
 template <typename T> // int key int index
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m256i vec0;
     __m256i vec1;
@@ -945,12 +945,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m256i p0;
     __m256i p1;
     uint8_t stride = (uint8_t)simd_width::AVX_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1147,7 +1147,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 
 template <typename T> // int key long index
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m256i vec0;
     __m256i vec1;
@@ -1157,12 +1157,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m256i pl1;
     __m256i ph1;
     uint8_t stride = (uint8_t)simd_width::AVX_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1370,17 +1370,17 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
 
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m256 vec0;
     __m256 vec1;
     
     uint8_t stride = (uint8_t)simd_width::AVX_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1549,7 +1549,7 @@ merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
 
 template <typename T> // float key int index
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m256 vec0;
     __m256 vec1;
@@ -1557,12 +1557,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m256i p0;
     __m256i p1;
     uint8_t stride = (uint8_t)simd_width::AVX_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1759,7 +1759,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 
 template <typename T> // float key long index
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m256 vec0;
     __m256 vec1;
@@ -1771,13 +1771,13 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m256i pl1;
     __m256i ph1;
     uint8_t stride = (uint8_t)simd_width::AVX_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
     // int buffer_ptr[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2000,17 +2000,17 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m256d vec0;
     __m256d vec1;
     
     uint8_t stride = (uint8_t)simd_width::AVX_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2179,7 +2179,7 @@ merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
 
 template <typename T> // double key int index
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m256d vec0;
     __m256d vec1;
@@ -2187,12 +2187,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m128i p0;
     __m128i p1;
     uint8_t stride = (uint8_t)simd_width::AVX_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2397,7 +2397,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 }
 template <typename T> // double key long index
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m256d vec0;
     __m256d vec1;
@@ -2405,12 +2405,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m256i p0;
     __m256i p1;
     uint8_t stride = (uint8_t)simd_width::AVX_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
